fix(prerender): Correctly parse broker descriptions
The regular expression used to extract broker descriptions from the `brokers.ts` file was too greedy, causing it to capture extraneous data. This change updates the regex to accurately capture only the content within the `description` field. This fix ensures that meta descriptions generated for broker pages are accurate and do not include unintended content, which was previously breaking the generation of meta tags.

Also adds FxPro broker to the `brokers.ts` data file.